introduce init.ts
